requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pydantic-settings==2.5.2
python-dotenv==1.0.1